requirements.txt
# Non-Airflow dependencies
google-auth==2.37.0  # Required for Google Cloud integration
pandas>=2.1.2       # Version required by airflow-providers-google
apache-airflow-providers-google>=10.23.0  # Latest version compatible with pandas>=2.1.2
ipykernel>=6.29.0   # For Jupyter notebook support
matplotlib==3.9.4   # For data visualization
ydata-profiling[notebook]==4.12.1  # For data profiling
openai==1.59.6      # For OpenAI API access
psycopg2-binary==2.9.9  # For PostgreSQL support
google-cloud-bigquery==3.27.0

